cp, install: use auto reflink by default
This patch is not for upstream consumption. It is only intended for
users of CoW filesystems that will always deduplicate things. For them
REFLINK_NEVER is about the same as REFLINK_AUTO with a week of delay.

The concern is usually minimized by the dedupe block size and the light
weight of the "deltas".

See: https://unix.stackexchange.com/questions/80351